# openapi.yaml
openapi: 3.1.0
info:
  title: Trading Engine API
  version: 1.0.0
  description: |
    REST + WebSocket API for a live trading engine ported from MQL/StereoTrader.

    ## Architecture
    The engine is **stateful** — all positions and orders live in a single in-memory
    instance per server process. State resets on restart.
    Calls are not isolated by user in the default paper-broker mode.

    ## Workflow — Getting Started

    The engine is **bar-driven**: you must push price bars before placing trades.
    A typical session looks like this:

    1. **Push bars** via `POST /bars` to establish a price reference
    2. **Open positions** via `POST /positions/long` or `POST /positions/short`
    3. **Set risk management** via `PUT /positions/{side}/sl-tp`
    4. **Place pending orders** via `POST /orders`
    5. **Monitor in real-time** via `GET /stream` (WebSocket)

    ```bash
    # 1. Feed the first bar
    curl -X POST http://localhost:3000/bars \
      -H 'Content-Type: application/json' \
      -d '{"bar":{"open":1.085,"high":1.087,"low":1.084,"close":1.086,"time":"2024-01-15T10:00:00Z"},"bars":[]}'

    # 2. Open a long position (1.0 lot)
    curl -X POST http://localhost:3000/positions/long \
      -H 'Content-Type: application/json' \
      -d '{"size":1.0}'

    # 3. Set a 50-point stop-loss
    curl -X PUT http://localhost:3000/positions/long/sl-tp \
      -H 'Content-Type: application/json' \
      -d '{"sl":50,"slActive":true}'
    ```

...

    ## Authentication
...
servers:
...
paths:
...
  /services/{id}:
    get:
      tags: [Services]
      summary: Get service detail
      operationId: getService
      description: |
        Returns full detail for a single service including the embedded
        `health` object with `status`, `lastHeartbeat`, `error`, and
        arbitrary `metadata` populated by the service implementation.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service identifier (as returned by `GET /services`)
          schema:
            type: string
          example: "broker:paper:primary"
      responses:
        '200':
          description: Service detail with health
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetail'
              example:
                id: "broker:paper:primary"
                kind: BROKER
                name: paper-primary
                health:
                  status: RUNNING
                  lastHeartbeat: "2026-03-07T12:00:00.000Z"
                  error: null
                  metadata:
                    connectedSince: "2026-03-07T11:00:00.000Z"
        '404':
          $ref: '#/components/responses/NotFound'
  /services/{id}/start:
    post:
      tags: [Services]
      summary: Start a service
      operationId: startService
      description: |
        Calls `svc.start()` on the named service and returns the updated
        `ServiceDetail`. If the service is already running the call is a
        no-op (implementation-defined).

        Returns `500` if `svc.start()` throws.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service identifier
          schema:
            type: string
          example: "data:mt5"
      responses:
        '200':
          description: Service started — returns updated detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Service start failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Connection refused: mt5://localhost:18812"
  /services/{id}/stop:
    post:
      tags: [Services]
      summary: Stop a service
      operationId: stopService
      description: |
        Calls `svc.stop()` on the named service and returns the updated
        `ServiceDetail`. If the service is already stopped the call is a
        no-op (implementation-defined).

        Returns `500` if `svc.stop()` throws.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service identifier
          schema:
            type: string
          example: "data:mt5"
      responses:
        '200':
          description: Service stopped — returns updated detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Service stop failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Timeout waiting for service to drain connections"
  /services/strategies/{id}/evaluate:
    post:
      tags: [Services]
      summary: On-demand strategy evaluation
      operationId: evaluateStrategy
      description: |
        Invokes `svc.evaluate()` on the named strategy service with a minimal
        context (empty bars, flat position). Returns the signal result
        (`BUY`, `SELL`, or `HOLD`).

        The strategy must be registered in the service registry with
        `kind: STRATEGY`. Returns `400` if the service is not a strategy.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Strategy service identifier
          schema:
            type: string
          example: strategy:candle-atr:eurusd-h1
      responses:
        '200':
          description: Strategy evaluation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  action:
                    type: string
                    enum: [BUY, SELL, HOLD]
              example:
                action: HOLD
        '400':
          description: Service is not a strategy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /services/screeners/{id}/scan:
    post:
      tags: [Services]
      summary: On-demand screener scan
      operationId: scanScreener
      description: |
        Invokes `svc.scan()` on the named screener service and returns all
        matches. The screener must be registered with `kind: SCREENER`.

        Returns `400` if the service is not a screener.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Screener service identifier
          schema:
            type: string
          example: screener:volume-breakout:majors
      responses:
        '200':
          description: Screener scan results
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items:
                      type: object
                      properties:
                        symbol:
                          type: string
                        matchType:
                          type: string
                        score:
                          type: number
                        metadata:
                          type: object
              example:
                matches:
                  - symbol: EURUSD
                    matchType: volume-breakout
                    score: 0.92
                    metadata: {}
        '400':
          description: Service is not a screener
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /services/risk/validate:
    post:
      tags: [Services]
      summary: On-demand risk validation
      operationId: validateRisk
      description: |
        Validates an order request against the risk manager's current limits
        (max open positions, max per symbol, max daily loss). Returns whether
        the order would be approved or rejected, along with the reason.

        Requires at least one `RISK_MANAGER` service to be registered.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol, direction, lots]
              properties:
                symbol:
                  type: string
                  example: EURUSD
                direction:
                  type: string
                  enum: [BUY, SELL]
                lots:
                  type: number
                  minimum: 0
                  example: 1.0
      responses:
        '200':
          description: Risk validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  approved:
                    type: boolean
                  reason:
                    type: string
              example:
                approved: true
                reason: All risk checks passed
        '404':
          description: No risk manager registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# ─────────────────────────────────────────────────────────────
# Components
# ─────────────────────────────────────────────────────────────
components:
...
  responses:
...
    NotFound:
      description: Resource not found — the ID/ticket does not exist or was already filled/cancelled
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: "Order abc not found" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string, description: "Human-readable error message describing what went wrong" }
...
    ServiceStatus:
      type: string
      enum: [STOPPED, STARTING, RUNNING, DEGRADED, STOPPING, ERROR]
      description: Current lifecycle state of a service
      example: RUNNING
    ServiceKind:
      type: string
      enum: [BROKER, DATA_PROVIDER, STRATEGY, SCREENER, ORDER_MANAGER, RISK_MANAGER]
      description: Functional category of the service
      example: BROKER
...
    ServiceHealth:
      type: object
      required: [status, lastHeartbeat, error, metadata]
      description: Real-time health snapshot for a service instance
      properties:
        status:
          $ref: '#/components/schemas/ServiceStatus'
        lastHeartbeat:
          type: string
          format: date-time
          nullable: true
          description: ISO-8601 timestamp of the last successful heartbeat, or `null` if never started
          example: "2026-03-07T12:00:00.000Z"
        error:
          type: string
          nullable: true
          description: Last error message if `status` is `ERROR`, otherwise `null`
          example: null
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary key/value data provided by the service implementation (e.g. connection stats, queue depths)
          example:
            connectedSince: "2026-03-07T11:00:00.000Z"
    ServiceDetail:
      type: object
      required: [id, kind, name, health]
      description: Full service detail including embedded health snapshot
      properties:
        id:
          type: string
          example: "broker:paper:primary"
        kind:
          $ref: '#/components/schemas/ServiceKind'
        name:
          type: string
          example: Paper Broker
        health:
          $ref: '#/components/schemas/ServiceHealth'
...
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        API key guard applied to all state-mutating routes (bars, positions,
        orders, scaled-orders, engine/ATR config) and skills routes.

        Set the `API_KEY` environment variable on the server. When `API_KEY` is
        **not set**, the pre-handler is a no-op — all endpoints are accessible
        without a key (useful for local development and tests).

        When `API_KEY` **is set**, every marked endpoint requires:
        ```
        x-api-key: <your-key>
        ```
        Missing or incorrect keys return `401 Unauthorized`.

        Note: OpenBB data endpoints (`/openbb/*`) use a separate
        `OPENBB_API_KEY` guard via the `?apiKey=` query parameter — they are
        not covered by this scheme.